backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
langchain==0.2.1
langchain-google-genai==1.0.3
google-generativeai==0.7.0
